Refactored methods in game_stats & Create team_utility module
