+++ lib/game_stats.rb
@@ -1,10 +1,12 @@
 require_relative './stats'
 require_relative './goals_utility'
 require_relative './season_utility'
+require_relative './team_utility'
 
 class GameStats < Stats
   include GoalsUtility
   include SeasonUtility
+  include TeamUtility
   
   def initialize(locations)
     super
@@ -21,12 +23,14 @@ class GameStats < Stats
   def percentage_home_wins
     total_of_home_games = 0
     wins_at_home = 0 
-    @game_teams.each do | k, v |
-      if k.hoa == "home"
-        total_of_home_games += 1
-      end
-      if k.hoa == "home" && k.result == "WIN"
-        wins_at_home += 1
+    hoa_all_game_teams.each do |hoa, game_teams_array|
+      game_teams_array.each do |game_team|
+        if hoa == "home"
+          total_of_home_games += 1
+        end
+        if hoa == "home" && game_team.result == "WIN"
+          wins_at_home += 1
+        end
       end
     end
     percent_win = ((wins_at_home / total_of_home_games.to_f)).round(2)
@@ -34,27 +38,31 @@ class GameStats < Stats
 
   def percentage_visitor_wins
     total_of_home_games = 0
-    losses_at_home = 0 
-    @game_teams.each do | k, v |
-      if k.hoa == "home"
-        total_of_home_games += 1
-      end
-      if k.hoa == "home" && k.result == "LOSS"
-        losses_at_home += 1
+    wins_at_home = 0 
+    hoa_all_game_teams.each do |hoa, game_teams_array|
+      game_teams_array.each do |game_team|
+        if hoa == "home"
+          total_of_home_games += 1
+        end
+        if hoa == "home" && game_team.result == "LOSS"
+          wins_at_home += 1
+        end
       end
     end
-    percent_loss = ((losses_at_home / total_of_home_games.to_f)).round(2)
+    percent_win = ((wins_at_home / total_of_home_games.to_f)).round(2)
   end
 
   def percentage_ties
     ties = 0 
-    total_of_games = @game_teams.count
-    @game_teams.each do | k, v |
-      if k.result == "TIE"
-        ties += 1
+    total_all_games = (hoa_all_game_teams["home"].count + hoa_all_game_teams["away"].count)
+    hoa_all_game_teams.each do |hoa, game_teams_array|
+      game_teams_array.each do |game_team|
+        if game_team.result == "TIE"
+          ties += 1
+        end
       end
     end
-    percent_ties = ((ties / total_of_games.to_f)).round(2)
+    percent_ties = ((ties / total_all_games.to_f)).round(2)
   end
 
   def count_of_games_by_season
